dashboard v2.2

+++ src/main/java/com/example/fumi_forte/dto/UsuarioReporteDto.java
@@ -1,7 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.example.fumi_forte.dto;
 
 import lombok.AllArgsConstructor;
@@ -12,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UsuarioReporteDto {
-    private int idUsuario;
+    private Long idUsuario;
     private String nombreCompleto;
     private String correo;
     private String estado;
